Moving `terminal.py` file (#2837)
* Move terminal file

* Update terminal_controller + paths

* Add new terminal.py

* Change constant : USER_CUSTOM_IMPORTS_DIRECTORY

* Changing constant : USER_PORTFOLIO_DATA_DIRECTORY

* Changing constant : ROUTINES_DIRECTORY

* Changing constant : DATA_SOURCES_DEFAULT_FILE

* Changing constant : PRESETS_DIRECTORY

* Update paths

* Update `terminal_controller` module

* Apply black

* Update worfklow + precommit hook

* Update workflow

* Tests : terminal_controller

* Tests : terminal_controller
