+++ openbb_terminal/portfolio/portfolio_analysis/pa_controller.py
@@ -9,7 +9,7 @@ from typing import List
 import pandas as pd
 
 from openbb_terminal import feature_flags as obbff
-from openbb_terminal.core.config.paths import PORTFOLIO_DATA_DIRECTORY
+from openbb_terminal.core.config.paths import USER_PORTFOLIO_DATA_DIRECTORY
 from openbb_terminal.custom_prompt_toolkit import NestedCompleter
 from openbb_terminal.decorators import log_start_end
 from openbb_terminal.menu import session
@@ -22,7 +22,7 @@ from openbb_terminal.rich_config import console
 
 logger = logging.getLogger(__name__)
 
-portfolios_path = PORTFOLIO_DATA_DIRECTORY / "portfolios"
+portfolios_path = USER_PORTFOLIO_DATA_DIRECTORY / "portfolios"
 port_types = [".csv", ".json", ".xlsx"]
 possible_paths = {
     portpath.name: portpath
