+++ openbb_terminal/stocks/insider/openinsider_model.py
@@ -10,7 +10,7 @@ import requests
 from bs4 import BeautifulSoup
 
 from openbb_terminal.decorators import log_start_end
-from openbb_terminal.core.config.paths import PRESETS_DIRECTORY
+from openbb_terminal.core.config.paths import USER_PRESETS_DIRECTORY
 from openbb_terminal.rich_config import console
 
 logger = logging.getLogger(__name__)
@@ -1109,7 +1109,7 @@ def get_preset_choices() -> Dict:
     filepath as value
     """
 
-    PRESETS_PATH = PRESETS_DIRECTORY / "stocks" / "insider"
+    PRESETS_PATH = USER_PRESETS_DIRECTORY / "stocks" / "insider"
     PRESETS_PATH_DEFAULT = Path(__file__).parent / "presets"
     preset_choices = {
         filepath.name: filepath
