+++ openbb_terminal/stocks/screener/finviz_model.py
@@ -13,12 +13,12 @@ from finvizfinance.screener import (
 )
 
 from openbb_terminal.decorators import log_start_end
-from openbb_terminal.core.config.paths import PRESETS_DIRECTORY
+from openbb_terminal.core.config.paths import USER_PRESETS_DIRECTORY
 from openbb_terminal.rich_config import console
 
 logger = logging.getLogger(__name__)
 
-PRESETS_PATH = PRESETS_DIRECTORY / "stocks" / "screener"
+PRESETS_PATH = USER_PRESETS_DIRECTORY / "stocks" / "screener"
 PRESETS_PATH_DEFAULT = Path(__file__).parent / "presets"
 preset_choices = {
     filepath.name: filepath
