+++ openbb_terminal/stocks/screener/screener_view.py
@@ -5,10 +5,10 @@ import configparser
 from pathlib import Path
 
 from openbb_terminal.rich_config import console
-from openbb_terminal.core.config.paths import PRESETS_DIRECTORY
+from openbb_terminal.core.config.paths import USER_PRESETS_DIRECTORY
 from openbb_terminal.stocks.screener import finviz_model
 
-PRESETS_PATH = PRESETS_DIRECTORY / "stocks" / "screener"
+PRESETS_PATH = USER_PRESETS_DIRECTORY / "stocks" / "screener"
 PRESETS_PATH_DEFAULT = Path(__file__).parent / "presets"
 preset_choices = {
     filepath.name: filepath
