+++ openbb_terminal/stocks/screener/yahoofinance_model.py
@@ -16,7 +16,7 @@ from sklearn.preprocessing import MinMaxScaler
 
 from openbb_terminal.decorators import log_start_end
 
-from openbb_terminal.core.config.paths import PRESETS_DIRECTORY
+from openbb_terminal.core.config.paths import USER_PRESETS_DIRECTORY
 from openbb_terminal.rich_config import console
 from openbb_terminal.stocks.screener import finviz_model
 
@@ -24,7 +24,7 @@ logger = logging.getLogger(__name__)
 
 register_matplotlib_converters()
 
-PRESETS_PATH = PRESETS_DIRECTORY / "stocks" / "screener"
+PRESETS_PATH = USER_PRESETS_DIRECTORY / "stocks" / "screener"
 PRESETS_PATH_DEFAULT = Path(__file__).parent / "presets"
 preset_choices = {
     filepath.name: filepath
